+++ src/aioway/fn/README.md
@@ -8,4 +8,4 @@ This means that `Fn`s would work with any valid `torch` operations and can be us
 This is done s.t. we can reuse what a lot of `torch` did without reinventing the wheel,
 which was attempted by koila and the previous versions.
 
-See `aioway.fate` for specifics.
+See `aioway.op` / `aioway.fate` / `aioway.might` for specifics.
