+++ src/aioway/fn/__init__.py
@@ -1,8 +1,7 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .fn import *
-from .nn import *
+from .modes import *
 from .op import *
 from .routers import *
-from .tensors import *
 from .tracking import *
