+++ src/aioway/fn/fn.py
@@ -3,7 +3,6 @@
 "Metadata for torch operators / functions."
 
 import abc
-import contextlib as ctxl
 import dataclasses as dcls
 import functools
 import logging
@@ -12,7 +11,7 @@ from collections import abc as cabc
 
 import torch
 
-from aioway._common import Stack, find_nested_tensors, render_fcall
+from aioway._common import find_nested_tensors, render_fcall
 
 __all__ = ["Fn", "TensorInput", "Thunk", "TorchThunk"]
 
@@ -194,75 +193,3 @@ class TorchThunk[T: cabc.Callable[..., typing.Any]](abc.ABC):
     def inputs(self):
         yield from find_nested_tensors(self.args)
         yield from find_nested_tensors(self.kwargs)
-
-
-@dcls.dataclass
-class OnOffCtx(abc.ABC):
-    """
-    `OnOffCtx` is a mixin class that gives the subclasses a toggle.
-    """
-
-    STACK: typing.ClassVar[OnOffStack[typing.Self]]
-    "The stack."
-
-    _: dcls.KW_ONLY
-
-    on: bool = True
-    "The toggle to control whether or not to run the current mode."
-
-    @abc.abstractmethod
-    def ctx(self) -> typing.ContextManager[typing.Self]:
-        raise NotImplementedError
-
-    @ctxl.contextmanager
-    def switch(self, on: bool, /):
-        "Switch to `on` in the scope (can be overwritten)."
-
-        before = self.on
-        self.on = on
-        try:
-            yield
-        finally:
-            self.on = before
-
-
-class OnOffStack[T: OnOffCtx](Stack[T]):
-    """
-    `OnOffStack` provides additional utilites to decide when to turn on or off.
-    """
-
-    @ctxl.contextmanager
-    def switch(self, on: bool | list[bool], /):
-        """
-        Temporarily set the `on` switch to the value given.
-        """
-
-        before = self.on
-        self.on = on
-
-        try:
-            yield
-        finally:
-            self.on = before
-
-    @property
-    def on(self) -> list[bool]:
-        "Get the on off values."
-
-        return [frame.on for frame in self]
-
-    @on.setter
-    def on(self, to: bool | list[bool]) -> None:
-        LOGGER.debug("Current stack %s", self)
-        LOGGER.debug("Setting to %s", to)
-
-        if isinstance(to, bool):
-            to = [to] * len(self)
-
-        if len(to) != len(self):
-            raise ValueError(f"Value {to=} should have equal length with {self=}.")
-
-        for frame, val in zip(self, to):
-            frame.on = val
-
-        LOGGER.debug("Status after setting %s", self)
